ODFTOOLKIT-475 validator: fix checking character content in foreign elements
... below text:p.

If the ODF element does not allow character content itself,
then for any child foreign element the character content should be
ignored.

See ODF 1.2 part 1 section 3.17 Foreign Elements and Attributes.

It's apparently not possible to ask the RelaxNG validator if an element
allows character content, so add a hard-coded list.