validator: add fallback validation of manifest.xml to work around OpenJDK bug
The problem is that ZipInputStream can't open the encrypted documents
that LO writes, you get an exception:
ZIP entry 'mimetype': only DEFLATED entries can have EXT descriptor

https://bugs.openjdk.java.net/browse/JDK-8143613

It's not possible to validate most of an encrypted document anyway, but it is
possible and very useful to validate the manifest.xml itself, because
(obviously) this is the only situation when it contains encryption related
markup.

The OdfPackage can't be used for this since it depends on the buggy
java.util.zip stuff.

So implement a fallback, when the expected ZipException arrives
OdfRootPackageValidator goes on to parse the ZIP central directory itself and
then reads META-INF/manifest.xml and validates only that.

Adapt ODFStreamValidator so that it copies the InputStream into a buffer
first; if OdfPackage.loadPackage is successful the buffer is discarded.

Add testcase validateEncryptedODF13WithInvalidOpenPGP.