OdfPackage: fix different handling of extra field using common compress
Apparently commons compress handles the Zip extra fields a bit
differently than java.util.zip.* does. The java.util.zip.ZipEntry, which
ZipArchiveEntry is a subclass of, states that getExtra() "returns the
extra field data for the entry, or null if none". While that seems to be
true for java.util.zip.*, commons compress apparently returns an emtpy
byte[] if the extra field is absent. Thus change the validation logic
for the mimetype to not only check for getExtra() being null but also
check if its size is > 0.