validator: auto-detect if MathML 1 DTD should be used
It's not very useful to have this be a configuration parameter passed in
for an entire ODF package; the MathML embedded objects in a package
may be different versions, so it's better to auto-detect the version.

Heuristic used is that if the 2nd line of the file has a DOCTYPE, then
use MathML 1 DTD, otherwise use MathML 2 schema.