ODFTOOLKIT-187: fix results of PuzzlePiece.extractPuzzlePieces()
This created nondeterministic and wrong results, somewhere in
reduceAttributes().

The reason appears to be that there are ordered containers like a
TreeSet<PuzzlePiece> in uniteDefinitionsWithEqualContent() involved
and the PuzzlePiece.compareTo() has a problem with overflow, so it
sometimes returns wrong results and then the TreeSet will be wrongly
sorted and it can't find some elements.

Just compare the integers properly and the result is deterministic.