odfdom: remove crufty mimetype code from OdfFileEntry
The "javax.activation" package was removed from JDK9, and is now
maintained by Eclipse, and is surely excessive for this functionality.

The fallback code here that is using a static map is mainly for file
types one would expect to be common on a 1990s Solaris workstation.

It turns out that there is a URLConnection.guessContentTypeFromName()
since forever so just use that.