odfdom: always use ZipFile instead of ZipArchiveInputStream
The documentation points out numerous pitfalls with
ZipArchiveInputStream, so don't use it in main code:

https://commons.apache.org/proper/commons-compress/zip.html#ZipArchiveInputStream_vs_ZipFile

The zip file was already read into a byte[] before, so just wrap that in
SeekableInMemoryByteChannel.

This compiles fine but somehow a too old commons-io is used to run unit
tests, so fix its version in the top-level pom.xml.